+++ src/ts/components/button.ts
@@ -2,6 +2,7 @@ import {ComponentConfig, Component} from './component';
 import {DOM} from '../dom';
 import {EventDispatcher, NoArgs, Event} from '../eventdispatcher';
 import { LocalizableText , i18n } from '../localization/i18n';
+import { Label, LabelConfig } from './label';
 
 /**
  * Configuration interface for a {@link Button} component.
@@ -22,12 +23,20 @@ export interface ButtonConfig extends ComponentConfig {
    * Default: false
    */
   acceptsTouchWithUiHidden?: boolean;
+
+  /**
+   * Specifies whether the text of the button should be shown when the button is hovered or focused.
+   *
+   * Default: false
+   */
+  showTextOnFocus?: boolean;
 }
 
 /**
  * A simple clickable button.
  */
 export class Button<Config extends ButtonConfig> extends Component<Config> {
+  protected textLabel: Label<LabelConfig>;
 
   private buttonEvents = {
     onClick: new EventDispatcher<Button<Config>, NoArgs>(),
@@ -41,7 +50,12 @@ export class Button<Config extends ButtonConfig> extends Component<Config> {
       role: 'button',
       tabIndex: 0,
       acceptsTouchWithUiHidden: false,
+      showTextOnFocus: false,
     } as Config, this.config);
+
+    if (this.config.showTextOnFocus) {
+      this.config.cssClasses = this.config.cssClasses.concat('ui-labeledbutton');
+    }
   }
 
   protected toDomElement(): DOM {
@@ -61,10 +75,21 @@ export class Button<Config extends ButtonConfig> extends Component<Config> {
       buttonElementAttributes['role'] = this.config.role;
     }
 
+    this.textLabel = new Label({
+      text: i18n.performLocalization(this.config.text),
+      for: this.config.id,
+      hidden: true,
+    });
+
     // Create the button element with the text label
-    let buttonElement = new DOM('button', buttonElementAttributes, this).append(new DOM('span', {
-      'class': this.prefixCss('label'),
-    }).html(i18n.performLocalization(this.config.text)));
+    let buttonElement = new DOM('button', buttonElementAttributes, this).append(this.textLabel.getDomElement());
+
+    if (this.config.showTextOnFocus) {
+      buttonElement.on('focusin', (e) => this.textLabel.show());
+      buttonElement.on('mouseenter', (e) => this.textLabel.show());
+      buttonElement.on('focusout', (e) => this.textLabel.hide());
+      buttonElement.on('mouseleave', (e) => this.textLabel.hide());
+    }
 
     // Listen for the click event on the button element and trigger the corresponding event on the button component
     buttonElement.on('click', () => {
@@ -79,11 +104,21 @@ export class Button<Config extends ButtonConfig> extends Component<Config> {
    * @param text the text to put into the label of the button
    */
   setText(text: LocalizableText): void {
-    this.getDomElement().find('.' + this.prefixCss('label')).html(i18n.performLocalization(text));
+    this.textLabel.setText(text);
   }
 
   protected onClickEvent() {
     this.buttonEvents.onClick.dispatch(this);
+  }
+
+  initialize(): void {
+    super.initialize();
+    this.textLabel.initialize();
+  }
+
+  release(): void {
+    super.release();
+    this.textLabel.release();
   }
 
   /**